layers are seperated according to clean architecture

+++ models/user.go
@@ -2,6 +2,6 @@ package models
 
 type User struct {
 	Token               string `json:"token" bson:"token"`
-	TelegramId          int    `json:"telegramId" bson:"telegramId"`
+	TelegramId          int64  `json:"telegramId" bson:"telegramId"`
 	TelegramDisplayName string `json:"telegramDisplayName" bson:"telegramDisplayName"`
 }
